+++ test/validators.js
@@ -1851,6 +1851,35 @@ describe('Validators', function () {
     });
   });
 
+  it('should validate ISSNs', function () {
+    test({
+      validator: 'isISSN',
+      valid: [
+        '0378-5955',
+        '0000-0000',
+        '2434-561X',
+        '2434-561x',
+      ],
+      invalid: [
+        '0378-5954',
+        '0000-0001',
+        '0378-123',
+        '037-1234',
+        '0',
+      ],
+    });
+    test({
+      validator: 'isISSN',
+      args: [{ case_sensitive: true }],
+      valid: [
+        '2434-561X',
+      ],
+      invalid: [
+        '2434-561x',
+      ],
+    });
+  });
+
   it('should validate JSON', function () {
     test({
       validator: 'isJSON',