+++ CHANGELOG.md
@@ -4,6 +4,8 @@
   ([#585](https://github.com/chriso/validator.js/pull/585))
 - Added support for greater or less than in `isFloat()`
   ([#544](https://github.com/chriso/validator.js/issues/544))
+- Added support for ISSN validation via `isISSN()`
+  ([#593](https://github.com/chriso/validator.js/pull/593))
 
 #### 6.0.0